add cancel function to windows version

+++ code/projects/TC_Tracker/core/include/multiArray.h
@@ -136,7 +136,7 @@ public:
 #ifdef _OPENMP
 #       pragma omp parallel for num_threads(threadNum)
 #endif
-        for (size_t i = 0; i < _times; ++i) {
+        for (int i = 0; i < _times; ++i) {
 //            avgValue += (*std::min_element(i*_rows*_columns + data.get(), (i+1)*_rows*_columns + data.get()));
             avgValues[i] = (*std::min_element(i*_rows*_columns + data.get(), (i+1)*_rows*_columns + data.get()));
         }
